Fix freeze when importing
